fix(lint): apply feedback
