+++ crates/biome_js_analyze/tests/specs/nursery/noReactForwardRef/valid.js
@@ -1,0 +1,17 @@
+/* should not generate diagnostics */
+
+const Component1 = ({ ref }) => {
+	return null;
+};
+
+const Component2 = ({ ref, ...props }) => {
+	return null;
+};
+
+const Component3 = (props) => {
+	return null;
+};
+
+const Component4 = ({ foo, bar }) => {
+	return null;
+};
